# ♻️ Refactor: Migrate randString to rand v2